+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.*;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.dto.Winners;
 import racingcar.domain.vo.Name;
 
 class RacingGameTest {
